requirements.txt
socket
logging
python-dotenv==0.19.0
emoji==1.5.0
python==3.8.12
pymongo==3.12.0
pandas